backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "摄像头场景大模型评测平台后端服务"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.0.0",
    "tos>=2.7.0",
    "volcengine-python-sdk[ark]>=1.0.0",
    "aiomysql>=0.2.0",
    "requests>=2.31.0",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
